refactor: extract service types into dedicated modules
