+++ src/services/types/courseTypes.ts
@@ -1,0 +1,21 @@
+/**
+ * Course Service Types
+ * Type definitions for course-related API operations
+ */
+
+export interface CourseEnrollment {
+  section?: string
+  studentSection?: string
+  courseSection?: string
+  [key: string]: unknown
+}
+
+export interface StudentCourseDto {
+  courseId: number
+  code: string
+  name: string
+  instructorName?: string
+  instructorUsername?: string
+  status: string
+  category?: string
+}
