+++ src/services/types/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Service Types Index
+ * Centralized exports for all service type definitions
+ */
+
+// Course types
+export * from './courseTypes'
+
+// Quiz types
+export * from './quizTypes'
